+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 * Unreleased
+    * **Breaking Change** Upgrade to AceTMI v0.5 which renames TM1637 classes
+      to be consistent with the TM1638 classes:
+        * `SimpleTmiInterface` to `SimpleTmi1637Interface`
+        * `SimpleTmiFastInterface` to `SimpleTmi1637FastInterface`
 * 0.10.0 (2022-02-02)
     * Add support for TM1638 LED modules using the `Tm1638Module`.
         * Depends on AceTMI library v0.4.